requirements.txt
android_backup==0.2.0
appdirs==1.4.4
attrs==25.4.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
construct==2.10.70
croniter==6.0.0
cryptography==46.0.5
defusedxml==0.7.1
Flask==3.1.3
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
micloud==0.6
netifaces==0.11.0
pycparser==3.0
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-miio==0.5.12
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
tqdm==4.67.3
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.6
zeroconf==0.148.0
